+++ api/src/controllers/shows/series.ts
@@ -2,7 +2,7 @@ import { and, eq, sql } from "drizzle-orm";
 import { Elysia, t } from "elysia";
 import { db } from "~/db";
 import { shows } from "~/db/schema";
-import { prefix } from "~/elysia";
+import { prefix } from "~/base";
 import { KError } from "~/models/error";
 import { madeInAbyss } from "~/models/examples";
 import { FullSerie, Serie, SerieTranslation } from "~/models/serie";
