+++ api/src/controllers/shows/shows.ts
@@ -13,7 +13,7 @@ import {
 } from "~/models/utils";
 import { desc } from "~/models/utils/descriptions";
 import { getShows, showFilters, showSort } from "./logic";
-import { prefix } from "~/elysia";
+import { prefix } from "~/base";
 
 export const showsH = new Elysia({ prefix: "/shows", tags: ["shows"] })
 	.model({
