+++ api/src/controllers/studios.ts
@@ -32,7 +32,7 @@ import {
 } from "~/models/utils";
 import { desc } from "~/models/utils/descriptions";
 import { getShows, showFilters, showSort } from "./shows/logic";
-import { prefix } from "~/elysia";
+import { prefix } from "~/base";
 
 const studioSort = Sort(
 	{
